Add ListBox Operations functionality
Added ListBox operations functionality to the project. The user can now insert an item to the ListBox, remove an item from the ListBox, clear all items from the ListBox, and search for a specific item in the ListBox. The functionality is implemented using the respective methods of the ListBox control.