Added Form1.vb file for printing functionality
This commit adds the Form1.vb file which contains the code for printing functionality. When the user clicks on the Button1, the PrintDocument1.Print() function is called to print the text entered in the TextBox1. The PrintDocument1_PrintPage event handler is used to define the layout of the printed document, including the font, margins, paper size, and rectangle. The commit includes the necessary code to ensure that the text is printed with high quality and is easy to read. This new functionality will be useful for users who need to print out important documents from this application.