Add basic drawing functionality to a Windows Form application
This commit adds basic drawing functionality to a Windows Form application. The application allows the user to draw shapes like rectangles, ellipses, lines, and freehand drawings. It also includes features like changing the background color of the paint area, changing the color of the shapes, and turning off the computer. The code has been tested and works as expected.